+++ source/command/WithdrawFundsCommand.cpp
@@ -1,4 +1,4 @@
 #include "../../include/command/WithdrawFundsCommand.h"
 
-WithdrawFundsCommand::WithdrawFundsCommand(int user_id, int quantity)
+WithdrawFundsCommand::WithdrawFundsCommand(int user_id, double quantity)
         : user_id_(user_id), quantity_(quantity) {}
